+++ src/drawer-00/story.tsx
@@ -2,7 +2,6 @@ import React, { Component } from 'react'
 import { text, withKnobs } from '@storybook/addon-knobs'
 import { storiesOf } from '@storybook/react'
 
-import { Drawer } from '../drawer'
 import { HamburgerButton } from '../hamburgerButton'
 import { BellIcon } from '../icon/bellIcon'
 import { BubbleIcon } from '../icon/bubbleIcon'
@@ -11,6 +10,7 @@ import { NewspaperIcon } from '../icon/newspaperIcon'
 import { TicketIcon } from '../icon/ticketIcon'
 import { ItemChoice, ItemChoiceStatus } from '../itemChoice'
 import { Menu } from '../menu'
+import { Drawer } from '.'
 
 const stories = storiesOf('Widgets|Drawer', module)
 stories.addDecorator(withKnobs)
